requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.27.0
attrs==24.3.0
certifi==2024.12.14
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.0
segno==1.6.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
yarl==1.18.3
